Extracted comment leveler.
